Adding a new experimental (and untested) SceneGraph helper DLL, which uses an object node graph API to manage and clone interdependent resources from packages. In time, hopefully this DLL could become a full object cloner.
The basic concept is that Nodes store resource's data and identifying key, while connections between them make changes to their parent node's data when their child node's identifier is changed.
KinHelpers are responsible for finding next of kin resources that are related to the parent resource purely by some common data shared by identifying keys.
The central ResourceGraph class is responsible for containing and managing instances of all three of the above classes, utilizing the FileTable to gather referenced and kindred resources.
The ResourceNodeDealer class is responsible for locating the correct nodes to contain specific types of resources that can have specialized behaviors to filter which referenced resources are "slurped" from the resource data and what kindred resources are located, if any, based on provided constraints.